Revamp with MySQL integrate

+++ src/main/java/com/thanu/work/todo2023/repository/ToDoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ToDoRepository extends CrudRepository<Todo,Long> {
+public interface ToDoRepository extends CrudRepository<Todo,Integer> {
     List<Todo> findAllToDosByUser(User user);
 
     long countToDosByUser(User user);
